Avoid useless warning from Qt
When cloning InsetExternal, the tempname_ member of InsetExternalParams
has yet to be initialized. So, trying to remove it causes the warning
"QFile::remove: Empty or null file name".

+++ src/insets/InsetExternal.cpp
@@ -99,7 +99,8 @@ TempName::~TempName()
 TempName & TempName::operator=(TempName const & other)
 {
 	if (this != &other) {
-		tempname_.removeFile();
+		if (!tempname_.empty())
+			tempname_.removeFile();
 		support::TempFile f("lyxextXXXXXX.tmp");
 		f.setAutoRemove(false);
 		tempname_ = f.name();
